officeShare_MongoDB/requirements.txt
motor==3.6.0
python-socketio==5.11.4
tornado==6.4.1
